ePay/server/requirements.txt
flask==3.0.3
stripe==10.12.0
google-auth==2.34.0
google-api-python-client==2.143.0
python-dotenv==1.0.1
pytest==8.3.3